backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
python-multipart
pillow
requests